build : deploy only on the production branch

+++ deploy.sh
@@ -1,21 +1,27 @@
-# Re-initialize dependencies and last exports/builds
-rm -rf node_modules/
-rm -rf .next/
-rm -rf out/
+BRANCH_NAME=$(git rev-parse --abbrev-ref HEAD)
 
-# Install dependencies, build and export (to generate `out/` folder)
-npm install
-npm run build
-npm run export
+if [ BRANCH_NAME == 'production' ]; then
+    # Re-initialize dependencies and last exports/builds
+    rm -rf node_modules/
+    rm -rf .next/
+    rm -rf out/
 
-# Add files that must be in the root of the `master` branch
-cd out/
-touch .nojekyll
-cp ../README.md .
+    # Install dependencies, build and export (to generate `out/` folder)
+    npm install
+    npm run build
+    npm run export
 
-# Push `out/` folder content on the `master` branch
-git init
-git add .
-git commit -m 'build : deploy on github pages'
-git remote add origin git@github.com:fduquesne/fduquesne.github.io.git
-git push --force origin master:master
+    # Add some files that must be in the root of the `master` branch
+    cd out/
+    touch .nojekyll
+    cp ../README.md .
+
+    # Push `out/` folder content on the `master` branch
+    git init
+    git add .
+    git commit -m 'build : deploy on github pages'
+    git remote add origin git@github.com:fduquesne/fduquesne.github.io.git
+    git push --force origin master:master
+else
+    echo 'ERROR - You must be on the `production` branch.'
+fi
